# [IMP] *: improve settings performance