Improve the res.config.settings wizard performance, on the python side (focus on `onchange`, `create` and `execute` rpcs).

* Avoid modification of non modified settings
* Do not fetch unnecessary data:
  * Use cached methods to avoid verifying n times the existence of a given record
  * Do not compute requested defaults if not asked by the ORM.

Before this PR: on runbot with "all" modules:
onchange: 0.6 sec
create: 0.53 s
execute (call_button): 0.59 s

After:
onchange: 0.27 sec
create: 0.31 sec
execute: 0.28 sec
and up to 3* less queries

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr