@Feyensv 2021-11-16
@rco-odoo Let's merge it in master, 15.0 is behind us and we should avoid changes in stable versions unless really necessary.

This code stayed like this for x years, merging it in master is good already, if really necessary for a specific instance, the bugfix will backport the commits.

Do you mind giving the r+ ?